Miscellaneous unit test clean up
