+++ src/lib/constants.ts
@@ -1,0 +1,4 @@
+/**
+ * The base URL for the You Need A Budget (YNAB) API.
+ */
+export const ynabApiBaseUrl = 'https://api.youneedabudget.com/v1';
