+++ src/lib/ynab-api/call-with-token.ts
@@ -1,11 +1,12 @@
 import { cookies } from 'next/headers';
+import { ynabApiBaseUrl } from '@/lib/constants';
 
 export async function callWithToken(endpoint: string) {
   const cookieStore = cookies();
   const headers = {
     Authorization: `Bearer ${cookieStore.get('ynabToken')?.value}`,
   };
-  return fetch(`https://api.youneedabudget.com/v1/${endpoint}`, {
+  return fetch(`${ynabApiBaseUrl}/${endpoint}`, {
     headers: new Headers(headers),
   })
     .then((res: Response) => {
